move Generate Execute to orchestrator
