+++ orchestrator/build.gradle
@@ -30,6 +30,8 @@ dependencies {
     testCompile 'org.mockito:mockito-all:1.9.5'
     testCompile 'org.junit.jupiter:junit-jupiter-engine:5.3.1'
     testCompile group: 'junit', name: 'junit', version: '4.12'
+    
+    testCompile project(':generator').sourceSets.test.output 
 }
 
 test {
